+++ R/sortableTableOutput.R
@@ -53,7 +53,7 @@ sortableTableOutput <- function(outputId) {
       )
     )
   )
-  jqui_sortabled(
+  jqui_sortable(
     tag = shiny::tableOutput(outputId),
     options = list(items = "tbody tr", shiny = shinyopt)
   )