+++ R/sortableTabsetPanel.R
@@ -50,7 +50,7 @@ sortableTabsetPanel <- function(..., id = NULL, selected = NULL,
 
   shiny_opt <- list(order = list(sortcreate = func, sortupdate = func))
 
-  jqui_sortabled(
+  jqui_sortable(
     tag = shiny::tabsetPanel(
       ..., id = id, selected = selected,
       type = type, position = position